requirements.txt
scholarly==1.7.11
gradio==3.23.0
loguru==0.6.0
pyyaml==5.4.1
openai==0.27.3
requests==2.25.1
fastapi==0.95.0
uvicorn==0.21.1
librosa==0.10.0
azure-cognitiveservices-speech==1.27.0
edge-tts==6.1.3
python-multipart==0.0.6